organized gitlab auth routes and API functions
