--- .env.development
@@ -1,7 +1,0 @@
-DB_NAME=badging_dev
-DB_USER=root
-DB_PASSWORD=rootpwd
-DB_HOST=127.0.0.1
-DB_DIALECT=mysql
-PORT=8000
-RETURN_JSON_ON_LOGIN=
