+++ providers/index.js
@@ -1,0 +1,9 @@
+const { githubAuth, githubAuthCallback } = require("./github/auth.js");
+const { gitlabAuth, gitlabAuthCallback } = require("./gitlab/auth.js");
+
+module.exports = {
+  githubAuth,
+  githubAuthCallback,
+  gitlabAuth,
+  gitlabAuthCallback,
+};
